feat(audit): overnight codebase health and improvement audit
Track A: Updated @cloudflare/workers-types.
Track B: Refactored discover.ts, fixed magic numbers and improved type safety.
Track C: Added 100% test coverage for core NLQ components.
Track D: Added JSDoc for public config and validation exports.
Included audit artifacts in plans/jules-audit/.
