+++ plans/jules-audit/AUDIT_DEPS.md
@@ -1,15 +1,8 @@
-# AUDIT DEPS
-
-## Findings
-The following dependencies were identified as outdated or having safe upgrades:
+# Track A - Dependency Audit
 
 | package | current | available | risk | upgrade safe? |
 |---------|---------|-----------|------|---------------|
-| @cloudflare/workers-types | 4.20260516.1 | 4.20260517.1 | Low | Yes (Minor) |
+| @cloudflare/workers-types | 4.20260517.1 | 4.20260518.1 | Low | Yes |
 
-## Actions Taken
-- Updated `@cloudflare/workers-types` to `4.20260517.1`.
-- Verified compilation and tests with `./scripts/quality_gate.sh`.
-
-## Human Review Required
-- `zod`: current 3.25.76, latest 4.4.3. Major version upgrade required, potential breaking changes.
+## Human Review Required (Major Versions)
+- zod: 3.25.76 -> 4.4.3
