+++ plans/jules-audit/AUDIT_QUALITY.md
@@ -1,20 +1,19 @@
-# AUDIT QUALITY
+# Track B - Code Quality
 
-## Findings
-- **Production Logs**: Found `console.log` statements in `worker/pipeline/score.ts`, `worker/publish.ts`, and `worker/lib/circuit-breaker.ts`. These should be handled by structured logging or removed.
-- **Magic Numbers**: Found several magic numbers in `worker/pipeline/discover.ts` related to adaptive budgeting and context windowing.
-- **File Size**: No files exceeded the 500-line limit in this track's scope, though some in the repo do (human review required).
+## File Length Violations (> 500 lines)
+- worker/pipeline/discover.ts (552 lines) - Action: Refactor to reduce size.
 
-## Actions Taken
-- Removed `console.log` from:
-  - `worker/pipeline/score.ts`
-  - `worker/publish.ts`
-  - `worker/lib/circuit-breaker.ts`
-- Refactored `worker/pipeline/discover.ts` to use a `DISCOVERY_CONSTANTS` object for all magic numbers.
-- Verified changes with `npx vitest run tests/unit/discover.test.ts`.
+## Magic Numbers
+- worker/state-machine.ts:108: 300 (lock extension)
+- worker/state-machine.ts:190: 1000 (retry delay)
+- worker/pipeline/discover.ts:239: 3 (concurrency)
+- worker/pipeline/normalize.ts:166: 4, 50 (code length)
+- worker/pipeline/dedupe.ts:40, 41, 42: 25, 100, 500 (value tiers)
 
-## Human Review Required
-- Large files identified for potential refactoring:
-  - `worker/lib/research-agent/fetcher.ts` (977 lines)
-  - `worker/lib/d1/queries.ts` (974 lines)
-  - `worker/lib/validation/reward-scraper.ts` (763 lines)
+## Untyped 'any'
+- worker/lib/github/workflows.ts:23: workflow_runs: any[]
+- worker/lib/github/core.ts:249: author: any
+
+## TODO/FIXME/HACK
+- worker/config.ts:46: HACKERNEWS (HACK)
+- worker/routes/webhooks.ts:4: DEPRECATED
