+++ worker/config.ts
@@ -146,6 +146,10 @@ export const CONFIG = {
 // Default Source Registry
 // ============================================================================
 
+/**
+ * Default source registry for the Deal Discovery System.
+ * Contains seed sources with their URL patterns and extraction selectors.
+ */
 export const DEFAULT_SOURCES = [
   {
     domain: "trading212.com",
@@ -263,6 +267,9 @@ export const DEFAULT_SOURCES = [
 // Error Messages
 // ============================================================================
 
+/**
+ * Standard error messages used throughout the system.
+ */
 export const ERROR_MESSAGES = {
   LOCK_CONFLICT: "Another run is in progress (lock conflict)",
   HASH_MISMATCH: "Snapshot hash mismatch - possible concurrent modification",
@@ -277,6 +284,9 @@ export const ERROR_MESSAGES = {
 // Validation Gate Names
 // ============================================================================
 
+/**
+ * Ordered list of mandatory validation gates in the worker pipeline.
+ */
 export const VALIDATION_GATES = [
   "schema_validation",
   "normalization_verification",
@@ -289,4 +299,7 @@ export const VALIDATION_GATES = [
   "snapshot_hash_verification",
 ] as const;
 
+/**
+ * Type representing a valid validation gate name.
+ */
 export type ValidationGate = (typeof VALIDATION_GATES)[number];
