+++ worker/lib/github/workflows.ts
@@ -20,8 +20,10 @@ export async function getWorkflowRuns(
     );
     if (!response.ok)
       throw new Error(`Failed to get workflow runs: ${response.status}`);
-    const data = await safeResponseJson<{ workflow_runs: any[] }>(response);
-    return (data?.workflow_runs || []).map((run) => ({
+    const data = await safeResponseJson<{
+      workflow_runs: Record<string, unknown>[];
+    }>(response);
+    return (data?.workflow_runs || []).map((run: any) => ({
       id: run.id,
       name: run.name,
       head_sha: run.head_sha,
