+++ worker/state-machine.ts
@@ -22,9 +22,17 @@ import { runExpirationCheck } from "./lib/expiration-manager";
 import type { Env } from "./types";
 
 // ============================================================================
-// State Machine Implementation
+// Constants
 // ============================================================================
 
+const PIPELINE_CONSTANTS = {
+  LOCK_EXTENSION_SECONDS: 300,
+  RETRY_BACKOFF_MS: 1000,
+} as const;
+
+// ============================================================================
+// State Machine Implementation
+// ============================================================================
 type StateHandler = (
   ctx: PipelineContext,
   env: Env,
@@ -105,7 +113,11 @@ export async function executePipeline(env: Env): Promise<{
       try {
         // Extend lock for long operations
         if (["discover", "validate", "publish"].includes(currentPhase)) {
-          await extendLock(env, trace_id, 300);
+          await extendLock(
+            env,
+            trace_id,
+            PIPELINE_CONSTANTS.LOCK_EXTENSION_SECONDS,
+          );
         }
 
         // Execute phase and instrument duration (discovery, validation, publish etc)
@@ -187,7 +199,12 @@ export async function executePipeline(env: Env): Promise<{
           ctx.retry_count++;
           if (ctx.metrics) recordRetry(ctx.metrics);
           // Retry same phase with backoff
-          await new Promise((r) => setTimeout(r, 1000 * ctx.retry_count));
+          await new Promise((r) =>
+            setTimeout(
+              r,
+              PIPELINE_CONSTANTS.RETRY_BACKOFF_MS * ctx.retry_count,
+            ),
+          );
           continue;
         }
