+++ worker/validation/types.ts
@@ -1,5 +1,8 @@
 import { Deal, PipelineContext } from "../types";
 
+/**
+ * Result of the full validation pipeline for a batch of deals.
+ */
 export interface ValidationResult {
   valid: Deal[];
   invalid: Array<{ deal: Deal; reasons: string[] }>;
@@ -13,6 +16,9 @@ export interface ValidationResult {
   };
 }
 
+/**
+ * Result of a single validation gate for a specific deal.
+ */
 export interface GateResult {
   passed: boolean;
   reason?: string;
